Use `double` for all floating-point numbers

+++ MaterialColorUtilities/Blend/Blender.cs
@@ -5,8 +5,8 @@ namespace MaterialColorUtilities.Blend;
 
 public static class Blender
 {
-    private const float HARMONIZE_MAX_DEGREES = 15.0f;
-    private const float HARMONIZE_PERCENTAGE = 0.5f;
+    private const double HARMONIZE_MAX_DEGREES = 15.0;
+    private const double HARMONIZE_PERCENTAGE = 0.5;
 
     /**
      * Blend the design color's HCT hue towards the key color's HCT hue, in a way that leaves the
@@ -21,13 +21,12 @@ public static class Blender
     {
         Hct fromHct = Hct.FromInt(designColor);
         Hct toHct = Hct.FromInt(sourceColor);
-        float differenceDegrees = (float)MathUtils.DifferenceDegrees(fromHct.Hue, toHct.Hue);
-        float rotationDegrees = Math.Min(differenceDegrees * HARMONIZE_PERCENTAGE, HARMONIZE_MAX_DEGREES);
-        float outputHue =
-            (float)
-                MathUtils.SanitizeDegreesDouble(
-                    fromHct.Hue
-                        + rotationDegrees * RotationDirection(fromHct.Hue, toHct.Hue));
+        double differenceDegrees = MathUtils.DifferenceDegrees(fromHct.Hue, toHct.Hue);
+        double rotationDegrees = Math.Min(differenceDegrees * HARMONIZE_PERCENTAGE, HARMONIZE_MAX_DEGREES);
+        double outputHue =
+            MathUtils.SanitizeDegreesDouble(
+                fromHct.Hue
+                    + rotationDegrees * RotationDirection(fromHct.Hue, toHct.Hue));
         return Hct.From(outputHue, fromHct.Chroma, fromHct.Tone).ToInt();
     }
 
@@ -40,13 +39,12 @@ public static class Blender
      * @param amount how much blending to perform; 0.0 >= and <= 1.0
      * @return from, with a hue blended towards to. Chroma and tone are constant.
      */
-    public static int BlendHctHue(int from, int to, float amount)
+    public static int BlendHctHue(int from, int to, double amount)
     {
         int ucs = BlendCam16Ucs(from, to, amount);
         Cam16 ucsCam = Cam16.FromInt(ucs);
         Cam16 fromCam = Cam16.FromInt(from);
-        return Hct.From(ucsCam.Hue, fromCam.Chroma, (float)ColorUtils.LStarFromArgb(from))
-            .ToInt();
+        return Hct.From(ucsCam.Hue, fromCam.Chroma, ColorUtils.LStarFromArgb(from)).ToInt();
     }
 
     /**
@@ -57,22 +55,22 @@ public static class Blender
      * @param amount how much blending to perform; 0.0 >= and <= 1.0
      * @return from, blended towards to. Hue, chroma, and tone will change.
      */
-    public static int BlendCam16Ucs(int from, int to, float amount)
+    public static int BlendCam16Ucs(int from, int to, double amount)
     {
         Cam16 fromCam = Cam16.FromInt(from);
         Cam16 toCam = Cam16.FromInt(to);
 
-        float aJ = fromCam.JStar;
-        float aA = fromCam.AStar;
-        float aB = fromCam.BStar;
+        double aJ = fromCam.JStar;
+        double aA = fromCam.AStar;
+        double aB = fromCam.BStar;
 
-        float bJ = toCam.JStar;
-        float bA = toCam.AStar;
-        float bB = toCam.BStar;
+        double bJ = toCam.JStar;
+        double bA = toCam.AStar;
+        double bB = toCam.BStar;
 
-        float j = aJ + (bJ - aJ) * amount;
-        float a = aA + (bA - aA) * amount;
-        float b = aB + (bB - aB) * amount;
+        double j = aJ + (bJ - aJ) * amount;
+        double a = aA + (bA - aA) * amount;
+        double b = aB + (bB - aB) * amount;
 
         Cam16 blended = Cam16.FromUcs(j, a, b);
         return blended.GetInt();
@@ -86,15 +84,15 @@ public static class Blender
      * @return -1 if decreasing from leads to the shortest travel distance, 1 if increasing from leads
      *     to the shortest travel distance.
      */
-    private static float RotationDirection(float from, float to)
+    private static double RotationDirection(double from, double to)
     {
-        float a = to - from;
-        float b = to - from + 360.0f;
-        float c = to - from - 360.0f;
+        double a = to - from;
+        double b = to - from + 360.0;
+        double c = to - from - 360.0;
 
-        float aAbs = Math.Abs(a);
-        float bAbs = Math.Abs(b);
-        float cAbs = Math.Abs(c);
+        double aAbs = Math.Abs(a);
+        double bAbs = Math.Abs(b);
+        double cAbs = Math.Abs(c);
 
         if (aAbs <= bAbs && aAbs <= cAbs)
         {
